Change: "Add automatic exchange rates update".
This code establishes a connection with a database and uses the apilayer API to fetch the latest exchange rates relative to the United States Dollar (USD). It then updates or inserts the exchange rate data into the database at regular intervals. The `scheduleExchangeRatesDataInsertion` function allows scheduling the automatic update of exchange rates in the database at a specified time interval. This process ensures that exchange rate data is always up to date and available for use.